backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
numpy>=1.24.0
chromadb>=0.4.0
llama-index>=0.9.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-chroma
llama-index-llms-openai
llama-index-readers-slack